+++ hive_io/lib/src/object/hive_object_internal.dart
@@ -24,37 +24,7 @@ extension HiveObjectInternal on HiveObjectMixin {
 
   /// Not part of public API
   void dispose() {
-    for (final list in _hiveLists.keys) {
-      (list as HiveListImpl).invalidate();
-    }
-
-    _hiveLists.clear();
-
     _box = null;
     _key = null;
   }
-
-  /// Not part of public API
-  void linkHiveList(HiveList list) {
-    _requireInitialized();
-    _hiveLists[list] = (_hiveLists[list] ?? 0) + 1;
-  }
-
-  /// Not part of public API
-  void unlinkHiveList(HiveList list) {
-    final currentIndex = _hiveLists[list]!;
-    final newIndex = _hiveLists[list] = currentIndex - 1;
-    if (newIndex <= 0) {
-      _hiveLists.remove(list);
-    }
-  }
-
-  /// Not part of public API
-  bool isInHiveList(HiveList list) {
-    return _hiveLists.containsKey(list);
-  }
-
-  /// Not part of public API
-  @visibleForTesting
-  Map<HiveList, int> get debugHiveLists => _hiveLists;
 }